+++ index.php
@@ -7,13 +7,14 @@
 
 <?php get_header() ?>
 
-<?php if ( have_posts() ) : while ( have_posts() ) : the_post(); ?>
-
-	<h1><?php the_title(); ?></h1>	
-	<?php the_content(); ?>
-
-<?php endwhile; else: ?>
-	<p><?php _e('Sorry, no posts matched your criteria.'); ?></p>
+<?php if (have_posts()): ?>
+		<?php while (have_posts()): ?>
+			<?php the_post(); ?>
+			<h1><?php the_title(); ?></h1>	
+			<?php the_content(); ?>
+		<?php endwhile; ?>
+	<?php else: ?>
+		<p><?php _e('Sorry, no posts matched your criteria.'); ?></p>
 <?php endif; ?>
 
 <?php get_footer() ?>