+++ php/functions/misc.php
@@ -84,6 +84,41 @@
 	include_once(TCS_URI.'/php/templates/acf/services.php');
 
 
+	add_action('admin_menu', function() {
+		remove_menu_page('edit.php');
+	});
+	add_action( 'init', function() {
+		$labels = array(
+			'name'                => __('Posts', 'text-domain'),
+			'singular_name'       => __('Post', 'text-domain'),
+			'add_new'             => _x('Add New Post', 'text-domain', 'text-domain'),
+			'add_new_item'        => __('Add New Post', 'text-domain'),
+			'edit_item'           => __('Edit Post', 'text-domain'),
+			'new_item'            => __('New Post', 'text-domain'),
+			'view_item'           => __('View Post', 'text-domain'),
+			'search_items'        => __('Search Posts', 'text-domain'),
+			'not_found'           => __('No Posts found', 'text-domain'),
+			'not_found_in_trash'  => __('No Posts found in Trash', 'text-domain'),
+			'parent_item_colon'   => __('Parent Post:', 'text-domain'),
+			'menu_name'           => __('Posts', 'text-domain'),
+		);
+
+		register_post_type( 'post', array(
+			'labels' 			=> $labels,
+			'public' 			=> true,
+			'_builtin' 			=> false, 
+			'_edit_link' 		=> 'post.php?post=%d', 
+			'capability_type' 	=> 'post',
+			'map_meta_cap' 		=> true,
+			'hierarchical' 		=> false,
+			'rewrite' 			=> array( 'slug' => 'blog' ),
+			'query_var' 		=> false,
+			'supports' 			=> array( 'title', 'editor', 'author', 'thumbnail', 'excerpt', 'trackbacks', 'custom-fields', 'comments', 'revisions', 'post-formats' ),
+			'menu_position' 	=> 5,
+		));
+	}, 1);
+
+
 	/**
 	 * CREATE CUSTOM POST TYPES
 	 */
@@ -92,8 +127,8 @@
 		'singular'			=> 'Portfolio',
 		'plural'			=> 'Portfolio',
 		'slug'				=> POST_TYPE_PORTFOLIO
-	), array('has_archive' => false));
-	$portfolio->register_taxonomy('type');
+	), array('has_archive' => false, 'menu_position' => 5));
+	$portfolio->register_taxonomy('kind');
 	$portfolio->register_taxonomy('post_tag');
 
 ?>