+++ php/templates/page-portfolio.php
@@ -5,45 +5,50 @@
 
 ?>
 
+<?php
+
+	$cardSettings = array(
+		'containerClass'		=> 'col-md-6',
+		'showCardDetails'		=> false,
+		'showFooterShare'		=> false,
+		'footerShareSettings'	=> array('id' => get_the_id(), 'alignRight' => true),
+	);
+
+?>
+
 <main class="md-cards-holder row">
 	<?php
-		global $post;
-		//global $the_query;
 		
 		$args = array(
-			'numberposts' 	=> -1,
-			'post_type' 	=> POST_TYPE_PORTFOLIO,
-			'post_tatus'	=> 'publish',
+			'posts_per_page' 	=> -1,
+			'post_type' 		=> POST_TYPE_PORTFOLIO,
+			'post_tatus'		=> 'publish',
 		);
-		$the_query = new WP_Query($args);
+		query_posts($args);
 		
-		if ($the_query->have_posts()):
-			while ($the_query->have_posts()):
-				$the_query->the_post();
-				?>
-				<div class="md-card-holder col-md-6">
-					<div class="md-card md-shadow-2dp">
-						<div class="md-card-header">
-							<h2><a href="<?php the_permalink() ?>"><?php the_title() ?></a></h2>
-						</div>
-						<div class="md-card-body">
-							<p><?php the_post_thumbnail(getPhotoSize()) ?></p>
-							<?php the_excerpt() ?>
-						</div>
-						<div class="md-card-footer row">
-							<div class="col-md-6">
-								<a href="<?php the_permalink() ?>" target="_self" class="btn btn-primary">Details</a>
-								<?php if (get_field('portfolio-type') == PORTFOLIO_WEBSITE): ?><a href="<?php echo get_field('project-url') ?>" target="_blank" class="btn btn-primary">Live Version</a><?php endif; ?>
-							</div>
-							<div class="item-footer-social col-md-6">
-								<?php displayShareButtons(array('id' => get_the_id(), 'alignRight' => true)) ?>
-							</div>
-						</div>
-					</div>
-				</div>
-				<?php
+		if (have_posts()):
+			while (have_posts()):
+				the_post();
+
+				$cardSettings['buttons'] = array(
+					'details'		=> array(
+						'url'		=> get_the_permalink(),
+						'label'		=> __('Details'),
+						'target'	=> '_self',
+						'class'		=> 'btn btn-primary',
+					));
+
+				if (get_field('portfolio-type') == PORTFOLIO_WEBSITE):
+					$cardSettings['buttons']['liveVersion']	= array(
+						'url'		=> get_field('website-url'),
+						'label'		=> __('Live version'),
+						'target'	=> '_blank',
+						'class'		=> 'btn btn-primary',
+					);
+				endif;
+
+				displayCard($cardSettings);
 			endwhile;
-			rewind_posts();
 		endif;
 	?>
 <main>